+++ packages/moonstone/MoonstoneDecorator/MoonstoneDecorator.js
@@ -26,6 +26,7 @@ import css from './MoonstoneDecorator.less';
 const defaultConfig = {
 	i18n: true,
 	float: true,
+	noAutoFocus: false,
 	overlay: false,
 	ri: {
 		screenTypes
@@ -50,7 +51,7 @@ const defaultConfig = {
  * @public
  */
 const MoonstoneDecorator = hoc(defaultConfig, (config, Wrapped) => {
-	const {ri, i18n, spotlight, float, overlay, textSize} = config;
+	const {ri, i18n, spotlight, float, noAutoFocus, overlay, textSize} = config;
 
 	// Apply classes depending on screen type (overlay / fullscreen)
 	const bgClassName = 'enact-fit' + (overlay ? '' : ` ${css.bg}`);
@@ -67,7 +68,7 @@ const MoonstoneDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			)
 		);
 	}
-	if (spotlight) App = SpotlightRootDecorator(App);
+	if (spotlight) App = SpotlightRootDecorator({noAutoFocus}, App);
 	if (textSize) App = TextSizeDecorator(App);
 
 	// add webOS-specific key maps